+++ src/main/java/com/nhnacademy/bookstoreback/bookcart/domain/entity/BookCart.java
@@ -15,6 +15,10 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+/**
+ * @author 이경헌
+ * Redis 에 저장되는 도서 장바구니 엔티티입니다.
+ */
 @RedisHash(value = "bookCarts")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -39,7 +43,7 @@ public class BookCart {
 			.build();
 	}
 
-	public void updateBookQuantity(Long bookId, Long bookQuantity) {
+	public void updateBookQuantity(Long bookId, Integer bookQuantity) {
 		Optional<Book> bookOptional = books.stream()
 			.filter(book -> book.getBookId().equals(bookId))
 			.findFirst();
